[Editor] Improve layout on smaller screens
Still not adapted for smartphones without exernal keyboard, but should be
working for small laptop/desktop screens (800*600 is now mostly usable, and
changes are visile under 1455px width screens). Tab scrolling is also improved
as buttons are no longer going out of the screen when too much tabs are
present.